requirements.txt
# Existing requirements
gradio>=4.0.0
langchain>=0.1.0
langchain-community>=0.0.20
crewai>=0.1.0
streamlit>=1.30.0
pandas>=1.5.0
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.15.0
requests>=2.31.0

# New backend requirements
# PDF processing and text extraction
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Vector database and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Computer vision for attention monitoring
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Machine Learning and NLP
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.6.0

# Data processing and storage
# sqlite3 is built into Python, no need to install
jsonschema>=4.17.0

# Async support
# asyncio is built into Python, no need to install
aiohttp>=3.8.0

# Date and time utilities
python-dateutil>=2.8.0

# Text processing
textblob>=0.17.0
regex>=2023.6.0

# Progress bars and utilities
tqdm>=4.65.0
rich>=13.0.0

# Image processing (optional, for enhanced attention monitoring)
Pillow>=10.0.0

# Audio processing (optional, for voice features - install manually if needed)
# pyaudio>=0.2.11
# SpeechRecognition>=3.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0